+++ betfairlightweight/endpoints/keepalive.py
@@ -27,10 +27,10 @@ class KeepAlive(BaseEndpoint):
 
         :rtype: KeepAliveResource
         """
-        (response, elapsed_time) = self.request(session=session)
-        self.client.set_session_token(response.get("token"))
+        (response, response_json, elapsed_time) = self.request(session=session)
+        self.client.set_session_token(response_json.get("token"))
         return self.process_response(
-            response, KeepAliveResource, elapsed_time, lightweight
+            response, response_json, KeepAliveResource, elapsed_time, lightweight
         )
 
     def request(
@@ -48,13 +48,13 @@ class KeepAlive(BaseEndpoint):
 
         check_status_code(response)
         try:
-            response_data = json_loads(response.text)
+            response_json = json_loads(response.text)
         except ValueError:
             raise InvalidResponse(response.text)
 
         if self._error_handler:
-            self._error_handler(response_data)
-        return response_data, elapsed_time
+            self._error_handler(response_json)
+        return response, response_json, elapsed_time
 
     def _error_handler(
         self, response: dict, method: str = None, params: dict = None