+++ betfairlightweight/endpoints/login.py
@@ -27,9 +27,13 @@ class Login(BaseEndpoint):
 
         :rtype: LoginResource
         """
-        (response, elapsed_time) = self.request(self.url, session=session)
-        self.client.set_session_token(response.get("sessionToken"))
-        return self.process_response(response, LoginResource, elapsed_time, lightweight)
+        (response, response_json, elapsed_time) = self.request(
+            self.url, session=session
+        )
+        self.client.set_session_token(response_json.get("sessionToken"))
+        return self.process_response(
+            response, response_json, LoginResource, elapsed_time, lightweight
+        )
 
     def request(
         self, method: str = None, params: dict = None, session: requests.Session = None
@@ -51,13 +55,13 @@ class Login(BaseEndpoint):
 
         check_status_code(response)
         try:
-            response_data = json_loads(response.text)
+            response_json = json_loads(response.text)
         except ValueError:
             raise InvalidResponse(response.text)
 
         if self._error_handler:
-            self._error_handler(response_data)
-        return response_data, elapsed_time
+            self._error_handler(response_json)
+        return response, response_json, elapsed_time
 
     def _error_handler(
         self, response: dict, method: str = None, params: dict = None