+++ tests/unit/test_account.py
@@ -29,7 +29,7 @@ class AccountTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.account.Account.process_response")
     @mock.patch(
         "betfairlightweight.endpoints.account.Account.request",
-        return_value=(mock.Mock(), mock.Mock()),
+        return_value=(mock.Mock(), mock.Mock(), mock.Mock()),
     )
     def test_get_account_funds(self, mock_request, mock_process_response):
         self.account.get_account_funds()
@@ -42,7 +42,7 @@ class AccountTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.account.Account.process_response")
     @mock.patch(
         "betfairlightweight.endpoints.account.Account.request",
-        return_value=(mock.Mock(), mock.Mock()),
+        return_value=(mock.Mock(), mock.Mock(), mock.Mock()),
     )
     def test_get_account_details(self, mock_request, mock_process_response):
         self.account.get_account_details()
@@ -55,7 +55,7 @@ class AccountTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.account.Account.process_response")
     @mock.patch(
         "betfairlightweight.endpoints.account.Account.request",
-        return_value=(mock.Mock(), mock.Mock()),
+        return_value=(mock.Mock(), mock.Mock(), mock.Mock()),
     )
     def test_get_account_statement(self, mock_request, mock_process_response):
         self.account.get_account_statement()
@@ -70,7 +70,7 @@ class AccountTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.account.Account.process_response")
     @mock.patch(
         "betfairlightweight.endpoints.account.Account.request",
-        return_value=(mock.Mock(), mock.Mock()),
+        return_value=(mock.Mock(), mock.Mock(), mock.Mock()),
     )
     def test_list_currency_rates(self, mock_request, mock_process_response):
         self.account.list_currency_rates()