+++ tests/unit/test_scores.py
@@ -26,7 +26,7 @@ class ScoresTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.scores.Scores.request")
     def test_list_race_details(self, mock_response):
         mock = create_mock_json("tests/resources/list_race_details.json")
-        mock_response.return_value = (mock.json(), 1.5)
+        mock_response.return_value = (mock.Mock(), mock.json(), 1.5)
 
         response = self.scores.list_race_details()
         assert mock.json.call_count == 1
@@ -37,7 +37,7 @@ class ScoresTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.scores.Scores.request")
     def test_list_available_events(self, mock_response):
         mock = create_mock_json("tests/resources/availableevents.json")
-        mock_response.return_value = (mock.json(), 1.3)
+        mock_response.return_value = (mock.Mock(), mock.json(), 1.3)
 
         response = self.scores.list_available_events()
         assert mock.json.call_count == 1
@@ -49,7 +49,7 @@ class ScoresTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.scores.Scores.request")
     def test_list_scores(self, mock_response):
         mock = create_mock_json("tests/resources/score.json")
-        mock_response.return_value = (mock.json(), 1.3)
+        mock_response.return_value = (mock.Mock(), mock.json(), 1.3)
         mock_update_keys = mock.Mock()
 
         response = self.scores.list_scores(mock_update_keys)
@@ -62,7 +62,7 @@ class ScoresTest(unittest.TestCase):
     @mock.patch("betfairlightweight.endpoints.scores.Scores.request")
     def test_list_incidents(self, mock_response):
         mock = create_mock_json("tests/resources/incidents.json")
-        mock_response.return_value = (mock.json(), 1.3)
+        mock_response.return_value = (mock.Mock(), mock.json(), 1.3)
         mock_update_keys = mock.Mock()
 
         response = self.scores.list_incidents(mock_update_keys)